requirements.txt
# Core dependencies
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0
polars>=0.18.0

# ML/Stats
scikit-learn>=1.3.0
xgboost>=2.0.0
lightgbm>=4.0.0
statsmodels>=0.14.0
shap>=0.42.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Web scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
playwright>=1.40.0

# Configuration & Utils
pyyaml>=6.0
pydantic>=2.0.0
python-dotenv>=1.0.0
tqdm>=4.65.0

# Scheduling & Async
APScheduler>=3.10.0

# Dashboard
streamlit>=1.28.0
streamlit-autorefresh>=1.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Additional utilities
joblib>=1.3.0
scipy>=1.11.0
